Added check for authorized access to todos

+++ Pyto/Services/TodoList/TodoListService.cs
@@ -2,6 +2,7 @@
 using Pyto.Data.Todo;
 using Pyto.Data.Users;
 using Pyto.Models;
+using Pyto.Services.Exceptions;
 
 namespace Pyto.Services.TodoList;
 
@@ -48,21 +49,28 @@ public class TodoListService : ITodoListService
 
 	public Task DeleteAsync(Todo todo)
 	{
-		// todo: delete only user's todo
-		return todoRepository.RemoveAsync(todo);
+		return DeleteAsync(todo.Id);
 	}
 
 	public async Task<Todo> UpdateAsync(Todo todo)
 	{
-		// todo: update only user's todo
+		todo = await todoRepository.ReadAsync(todo.Id).ConfigureAwait(false);
+		if (todo.AuthorId != Author.Id)
+		{
+			throw new ForbiddenException();
+		}
 		todo = await todoRepository.UpdateAsync(todo).ConfigureAwait(false);
 		return todo;
 	}
 
 	public async Task DeleteAsync(Guid todoId)
 	{
-		// todo: delete only user's todo
 		var todo = await todoRepository.ReadAsync(todoId).ConfigureAwait(false);
+		if (todo.AuthorId != Author.Id)
+		{
+			throw new ForbiddenException();
+		}
+
 		await DeleteAsync(todo).ConfigureAwait(false);
 	}
 }
